@@ -7,6 +7,7 @@
 		"dev": "vite dev",
 		"build": "vite build",
 		"preview": "vite preview",
+		"analyze": "ANALYZE=true vite build",
 		"prepare": "svelte-kit sync || echo ''",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
@@ -21,6 +22,7 @@
 		"@playwright/test": "^1.49.1",
 		"@plone/blocks": "1.0.0-alpha.3",
 		"@sveltejs/adapter-auto": "^4.0.0",
+		"@sveltejs/adapter-node": "^5.4.0",
 		"@sveltejs/kit": "^2.16.0",
 		"@sveltejs/vite-plugin-svelte": "^5.0.0",
 		"@tailwindcss/forms": "^0.5.9",
@@ -31,12 +33,13 @@
 		"jsdom": "^26.0.0",
 		"prettier": "^3.4.2",
 		"prettier-plugin-svelte": "^3.3.3",
-		"svelte": "^5.0.0",
+		"svelte": "^5.46.1",
 		"svelte-check": "^4.0.0",
 		"svelte-preprocess-react": "^2.1.0",
 		"tailwindcss": "^4.0.0",
 		"typescript": "^5.0.0",
 		"vite": "^6.0.0",
+		"vite-bundle-analyzer": "^1.3.2",
 		"vitest": "^3.0.0"
 	},
 	"pnpm": {
@@ -50,6 +53,8 @@
 		"@tanstack/svelte-query": "^5.69.0",
 		"@tanstack/svelte-query-devtools": "^5.69.0",
 		"@tanstack/svelte-query-persist-client": "^5.69.0",
+		"embla-carousel": "^8.6.0",
+		"embla-carousel-svelte": "^8.6.0",
 		"react": "^19.0.0",
 		"react-dom": "^19.0.0",
 		"sass": "^1.86.0",
