@@ -1,27 +1,36 @@
 <script lang="ts">
-	import DefaultBlockView from '$lib/blocks/DefaultBlockView.svelte';
-	let { content, blocksConfig = {}, pathname, metadata = undefined } = $props();
+	import RenderBlock from '$lib/RenderBlock.svelte';
+	import type { ListingResponse } from '$lib/blocks/listing/types';
+	import type { BlockConfig } from '$lib/blocks/index';
+
+	let {
+		content,
+		blocksConfig = {} as Record<string, BlockConfig>,
+		pathname,
+		metadata = undefined,
+		listingData = {} as Record<string, ListingResponse>,
+		listingPages = {} as Record<string, number>,
+		paginatedBlockCount = 1
+	} = $props();
 </script>
 
 {#if content?.blocks_layout?.items}
 	{#each content.blocks_layout.items as block}
 		{@const blockData = content.blocks?.[block]}
 		{@const blockType = blockData?.['@type']}
-		{@const BlockComponent = blocksConfig[blockType]?.view || DefaultBlockView}
-		<!-- {@debug blockData} -->
-		<div class="VoltoBlock">
-			{#if BlockComponent === DefaultBlockView}
-				<p>block: {block} [{blockType}]</p>
-			{/if}
-			<BlockComponent
-				key={block}
-				id={block}
-				data={blockData}
-				{metadata}
-				properties={content}
-				path={pathname}
-				{blocksConfig}
-			/>
-		</div>
+		{@const blockConfig = blocksConfig[blockType]}
+		<RenderBlock
+			{block}
+			{blockData}
+			{blockType}
+			{blockConfig}
+			{content}
+			{pathname}
+			{metadata}
+			{listingData}
+			{listingPages}
+			{paginatedBlockCount}
+			{blocksConfig}
+		/>
 	{/each}
 {/if}
