@@ -2,36 +2,30 @@
 	/**
 	 * ResponsiveImage component for rendering Plone images with srcset support.
 	 * Consolidates image rendering logic from ImageBlockView and TeaserBlockView.
+	 *
+	 * Width and height props are required for CLS (Cumulative Layout Shift) prevention.
+	 * These should always be set from Plone's image_scales data.
 	 */
 
 	interface ScaleEntry {
-		/** Download path for this scale */
 		download: string;
-		/** Width of this scale in pixels */
 		width: number;
-		/** Height of this scale in pixels (optional) */
 		height?: number;
 	}
 
 	interface Props {
-		/** Plone image_scales data containing available image sizes */
 		scales: Record<string, ScaleEntry> | undefined;
-		/** Base URL for constructing full image paths */
 		baseUrl: string;
-		/** Alt text for accessibility (required for screen readers) */
 		alt: string;
-		/** Direct src URL for the img element fallback */
 		src?: string;
-		/** Responsive sizing hints for browser optimization */
 		sizes?: string;
-		/** Image width for CLS prevention */
 		width?: number;
-		/** Image height for CLS prevention */
 		height?: number;
-		/** Loading strategy: 'lazy' (default) or 'eager' */
 		loading?: 'lazy' | 'eager';
-		/** Custom CSS class for the img element */
+		fetchpriority?: 'high' | 'low' | 'auto';
+		decoding?: 'async' | 'sync' | 'auto';
 		class?: string;
+		onload?: () => void;
 	}
 
 	let {
@@ -43,13 +37,14 @@
 		width,
 		height,
 		loading = 'lazy',
-		class: className
+		fetchpriority = 'auto',
+		decoding = 'async',
+		class: className,
+		onload
 	}: Props = $props();
 
-	/**
-	 * Generates srcset string from Plone image_scales data.
-	 * Format: "{baseUrl}/{download} {width}w, ..."
-	 */
+	let loaded = $state(false);
+
 	const srcset = $derived(() => {
 		if (!scales || Object.keys(scales).length === 0) {
 			return '';
@@ -64,16 +59,63 @@
 
 		return srcsetEntries.join(', ');
 	});
+
+	const containerStyle = $derived(() => {
+		const styles: string[] = [];
+
+		if (width && height) {
+			styles.push(`--aspect-ratio: ${width} / ${height}`);
+		}
+
+		styles.push('--image-placeholder-bg: #e5e5e5');
+
+		return styles.join('; ');
+	});
+
+	function handleLoad() {
+		loaded = true;
+		onload?.();
+	}
 </script>
 
-<picture>
-	<source srcset={srcset()} {sizes} />
-	<img
-		{src}
-		{alt}
-		width={width}
-		height={height}
-		{loading}
-		class={className}
-	/>
-</picture>
+<div
+	class="responsive-image-container"
+	class:loaded
+	style={containerStyle()}
+>
+	<picture>
+		<source srcset={srcset()} {sizes} />
+		<img
+			{src}
+			{alt}
+			{width}
+			{height}
+			{loading}
+			{fetchpriority}
+			{decoding}
+			class={className}
+			onload={handleLoad}
+		/>
+	</picture>
+</div>
+
+<style>
+	.responsive-image-container {
+		background-color: var(--image-placeholder-bg, #e5e5e5);
+		aspect-ratio: var(--aspect-ratio);
+	}
+
+	.responsive-image-container.loaded {
+		background-color: transparent;
+	}
+
+	.responsive-image-container picture {
+		display: block;
+	}
+
+	.responsive-image-container img {
+		display: block;
+		width: 100%;
+		height: auto;
+	}
+</style>
