@@ -1,19 +1,78 @@
-<script>
-	import Picture from '$lib/plone/Picture.svelte';
-	export let item = {};
-	const defaultScale = 'preview';
+<script lang="ts">
+	import ResponsiveImage from '$lib/components/ResponsiveImage.svelte';
 
-	// function getImageConfig(item, fieldName) {
-	// 	if (!item.image_scales) {
-	// 		return;
-	// 	}
-	// 	const imageConfig = item.image_scales.hasOwnProperty(fieldName)
-	// 		? item.image_scales[fieldName][0]
-	// 		: null;
-	// 	return imageConfig;
-	// }
+	interface ImageScale {
+		download: string;
+		width: number;
+		height?: number;
+	}
 
-	// const leadImage = getImageConfig(item, 'image');
+	interface ImageData {
+		download?: string;
+		width?: number;
+		height?: number;
+		scales?: Record<string, ImageScale>;
+		alt?: string;
+	}
+
+	interface ListingItemData {
+		href?: string;
+		title?: string;
+		description?: string;
+		image_field?: string;
+		image_scales?: Record<string, ImageData[]>;
+	}
+
+	export let item: ListingItemData = {};
+	const defaultFieldName = 'image';
+
+	function getImageData(itemData: ListingItemData): ImageData | null {
+		if (!itemData.image_scales) return null;
+		const fieldName = itemData.image_field || defaultFieldName;
+		const fieldData = itemData.image_scales[fieldName];
+		if (fieldData?.[0]) {
+			return fieldData[0];
+		}
+		return null;
+	}
+
+	function getImageScales(itemData: ListingItemData): Record<string, ImageScale> | undefined {
+		const imgData = getImageData(itemData);
+		return imgData?.scales;
+	}
+
+	function getImageBaseUrl(itemData: ListingItemData): string {
+		return itemData.href || '';
+	}
+
+	function getImageSrc(itemData: ListingItemData): string {
+		const imgData = getImageData(itemData);
+		if (!imgData?.download) return '';
+		const baseUrl = getImageBaseUrl(itemData);
+		return `${baseUrl}/${imgData.download}`;
+	}
+
+	function getImageAlt(itemData: ListingItemData): string {
+		const imgData = getImageData(itemData);
+		return imgData?.alt || itemData.title || '';
+	}
+
+	function getImageWidth(itemData: ListingItemData): number | undefined {
+		const imgData = getImageData(itemData);
+		return imgData?.width;
+	}
+
+	function getImageHeight(itemData: ListingItemData): number | undefined {
+		const imgData = getImageData(itemData);
+		return imgData?.height;
+	}
+
+	const scales = getImageScales(item);
+	const baseUrl = getImageBaseUrl(item);
+	const src = getImageSrc(item);
+	const alt = getImageAlt(item);
+	const width = getImageWidth(item);
+	const height = getImageHeight(item);
 </script>
 
 <div class="listingItem">
@@ -22,9 +81,17 @@
 		<p>{item.description}</p>
 	</div>
 	<div class="imageWrapper">
-		<!-- {#if leadImage} -->
-		<Picture catalogItem={item} {defaultScale} />
-		<!-- {/if} -->
+		{#if scales}
+			<ResponsiveImage
+				{scales}
+				{baseUrl}
+				{alt}
+				{src}
+				{width}
+				{height}
+				sizes="(max-width: 768px) 100vw, 400px"
+			/>
+		{/if}
 	</div>
 </div>
 
@@ -39,8 +106,6 @@
 	}
 	.textWrapper {
 		flex: 1 2 300px;
-		// flex-shrink: 2;
-		// flex-grow: 1;
 		padding-right: 1em;
 	}
 	.imageWrapper {
