@@ -1,12 +1,80 @@
+<!--
+@deprecated This component is deprecated. Use ResponsiveImage instead.
+
+Migration Guide:
+================
+Replace Picture.svelte with ResponsiveImage from '$lib/components/ResponsiveImage.svelte'.
+
+Props mapping:
+- catalogItem/objectItem -> Extract scales from image_scales and pass to `scales` prop
+- catalogItem.href -> `baseUrl` prop
+- fieldName -> Use to extract the correct field from image_scales
+- defaultScale -> Not needed, ResponsiveImage uses all available scales
+
+Example migration:
+==================
+
+Before (Picture.svelte):
+```svelte
+<Picture catalogItem={item} fieldName="image" />
+```
+
+After (ResponsiveImage):
+```svelte
 <script>
-	import { PUBLIC_FRONTEND_BASE_URL } from '$env/static/public';
+import ResponsiveImage from '$lib/components/ResponsiveImage.svelte';
+
+// Extract image data from catalog item
+const imageData = item.image_scales?.image?.[0];
+const scales = imageData?.scales;
+const baseUrl = item.href || '';
+const src = imageData?.download ? `${baseUrl}/${imageData.download}` : '';
+</script>
+
+<ResponsiveImage
+  {scales}
+  {baseUrl}
+  {src}
+  alt=""
+  width={imageData?.width}
+  height={imageData?.height}
+/>
+```
+
+ResponsiveImage benefits:
+- No /renderimg route dependency (uses direct URLs)
+- Supports fetchpriority prop for LCP optimization
+- CSS aspect-ratio placeholder for CLS prevention
+- Supports onload handler for fade-in transitions
+-->
+<script lang="ts">
+	interface Scale {
+		download: string;
+		width: number;
+		height?: number;
+	}
+
+	interface ImageConfig {
+		scales?: Record<string, Scale>;
+	}
+
+	interface CatalogItem {
+		href?: string;
+		image_scales?: Record<string, ImageConfig[]>;
+	}
+
+	interface ObjectItem {
+		href?: string;
+		[key: string]: unknown;
+	}
+
 	export let defaultScale = 'preview';
 	export let onlyDefaultScale = false;
 	export let fieldName = 'image';
-	export let catalogItem;
-	export let objectItem;
+	export let catalogItem: CatalogItem | undefined = undefined;
+	export let objectItem: ObjectItem | undefined = undefined;
 
-	function getImageConfigFromCatalogItem(item, fieldName) {
+	function getImageConfigFromCatalogItem(item: CatalogItem, fieldName: string): ImageConfig {
 		if (!item.image_scales) {
 			return {};
 		}
@@ -16,14 +84,12 @@
 		return imageConfig;
 	}
 
-	function getImageConfigFromObjectItem(item, fieldName) {
-		// console.log(item)
-		const imageConfig = item.hasOwnProperty(fieldName) ? item[fieldName] : {};
-		// console.log(imageConfig)
+	function getImageConfigFromObjectItem(item: ObjectItem, fieldName: string): ImageConfig {
+		const imageConfig = item.hasOwnProperty(fieldName) ? (item[fieldName] as ImageConfig) : {};
 		return imageConfig;
 	}
 
-	function generateSrcSet(imageConfig, basePath) {
+	function generateSrcSet(imageConfig: ImageConfig, basePath: string): string {
 		if (!imageConfig) {
 			return '';
 		}
@@ -35,25 +101,25 @@
 			if (srcSet) {
 				srcSet = srcSet + ',';
 			}
-			let srcEntry;
-			srcEntry = `${PUBLIC_FRONTEND_BASE_URL}/renderimg${basePath}/${scales[s].download} ${scales[s].width}w`;
+			let srcEntry: string;
+			// Use direct paths instead of /renderimg route
+			srcEntry = `${basePath}/${scales[s].download} ${scales[s].width}w`;
 			srcSet = srcSet + srcEntry;
-			// no download needed for objectItem ;) ^^
 		}
 		return srcSet;
 	}
 
-	let imageConfig = {};
+	let imageConfig: ImageConfig = {};
 	let basePath = '';
 	if (catalogItem) {
 		imageConfig = getImageConfigFromCatalogItem(catalogItem, fieldName);
-		basePath = catalogItem['href'];
+		basePath = catalogItem['href'] || '';
 	} else if (objectItem) {
 		imageConfig = getImageConfigFromObjectItem(objectItem, fieldName);
-		basePath = objectItem['href'];
+		basePath = objectItem['href'] || '';
 	}
 
-	const scales = imageConfig.hasOwnProperty('scales') ? imageConfig.scales : {};
+	const scales: Record<string, Scale> = imageConfig.hasOwnProperty('scales') && imageConfig.scales ? imageConfig.scales : {};
 
 	const scale = scales.hasOwnProperty(defaultScale) ? scales[defaultScale] : scales['preview'];
 	const srcset = generateSrcSet(imageConfig, basePath);
@@ -66,7 +132,7 @@
 			alt=""
 			class=""
 			loading="lazy"
-			src="{PUBLIC_FRONTEND_BASE_URL}/renderimg{basePath}/{scale.download}"
+			src="{basePath}/{scale.download}"
 			width={scale.width}
 			height={scale.height}
 		/>
